Migrate seeds from size_t to uint32_t

+++ src/fheroes2/kingdom/week.h
@@ -76,7 +76,7 @@ struct Week
 
     const char * GetName() const;
 
-    static Week RandomWeek( const World & world, const bool isNewMonth, const size_t weekSeed );
+    static Week RandomWeek( const World & world, const bool isNewMonth, const uint32_t weekSeed );
 
     friend StreamBase & operator>>( StreamBase & stream, Week & week );
